User suspension has been added.

+++ client/src/pages/Admin/Admin.tsx
@@ -22,9 +22,10 @@ function Admin() {
   const { token, logout } = AuthentificationHook();
   // navigate object to handle navigation after request, in this case if token is invalid go back to login
   const [modalBool, setModalBool] = useState<boolean>(false);
-  const [status, setStatus] = useState<string>("")
-  const [userId, setUserId] = useState<string>("")
-  const [questionId, setQuestionId] = useState<string>("")
+  const [suspend, setSuspend] = useState<boolean | null>();
+  const [status, setStatus] = useState<string | null>("")
+  const [userId, setUserId] = useState<string | null>("")
+  const [questionId, setQuestionId] = useState<string | null>("")
   const navigate = useNavigate();
 
   // useEffect to render on empty array list, page loading, to fill out questions
@@ -57,7 +58,7 @@ function Admin() {
       getUsersStats();
   }, [token, navigate, logout]);
 
-
+  console.log(users)
   // statusUpdate function that sends request to statusUpdate endpoint for approval or denial
   const statusUpdate = async (questionId: string, newStatus: string) => {
     try {
@@ -65,7 +66,7 @@ function Admin() {
       
       setQuestions((prev) =>
           prev.filter((question) => question.questionId !== questionId)
-        );
+      );
     }
     catch (err) {
       console.error(`Error updating question ${questionId} to ${newStatus}. Error: ${err}`)
@@ -87,7 +88,9 @@ function Admin() {
   }
 
   // bool function for modal
-  const handleModal = async (bool: boolean, userId?: string, questionId?: string, status?: string) => {
+  const handleModal = async (args: {bool: boolean, userId?: string, questionId?: string, status?: string, suspend?: boolean}) => {
+    const { bool, userId, questionId, status, suspend } = args;
+
     setModalBool(bool);
     if (userId){
       setUserId(userId)
@@ -97,6 +100,24 @@ function Admin() {
     }
     if (status){
       setStatus(status)
+    }
+    if (suspend !== undefined && suspend !== null){
+      setSuspend(suspend)
+    }
+  }
+
+  // function to update user's suspension
+  const handleSuspend = async (userId: string, suspend: boolean) => {
+    try {
+      const res = await adminService.updateUserSuspend(userId, suspend)
+      setUsers(prev =>
+        prev.map(user =>
+          user.userId === userId ? { ...user, suspended: res.suspend } : user
+      )
+    );
+    }
+    catch (err){
+      console.error(`Error updating user suspension. ${err}`)
     }
   }
 
@@ -157,6 +178,7 @@ function Admin() {
                         easy_count={user.easy_count}
                         med_count={user.med_count}
                         hard_count={user.hard_count}
+                        suspended={Boolean(user.suspended)}
                         handleModal={handleModal}
                       />
                     ))}
@@ -171,25 +193,39 @@ function Admin() {
               <div className="modal-dialog">
                   <div className="modal-content">
                       <div className="modal-header">
-                          <h5 className="modal-title">
-                            Are you sure you want to{" "}
-                            {tab === "questions" ? (status === "approved" ? "approve" : "deny") : "delete"}{" "}
-                            this {tab === "questions" ? "question" : "user"}?
-                          </h5>
+                        <h5 className="modal-title">
+                          Are you sure you want to{" "}
+                          {tab === "questions"
+                            ? status === "approved"
+                              ? "approve"
+                              : "deny"
+                            : suspend === true
+                              ? "suspend"
+                              : suspend === false
+                                ? "unsuspend"
+                                : "delete"}{" "}
+                          this {tab === "questions" ? "question" : "user"}?
+                        </h5>
                       </div>
                       <div className="modal-footer">
-                        <button type="button" className="btn btn-secondary" data-bs-dismiss="modal" onClick={() => handleModal(false)}> No </button>
+                        <button type="button" className="btn btn-secondary" data-bs-dismiss="modal" onClick={() => handleModal({bool: false})}> No </button>
                         <button type="button" className="btn btn-danger" onClick={() => {
-                          if (tab === "questions" && status){
+                          if (tab === "questions" && status && questionId){
                             statusUpdate(questionId, status)
                           }
-                          else if (tab === "users"){
-                            deleteUser(userId)
+                          else if (tab === "users" && userId){
+                            if (suspend != null){
+                              handleSuspend(userId, suspend)
+                            }
+                            else{
+                              deleteUser(userId)
+                            }
                           }
                           setModalBool(false)
-                          setStatus("")
-                          setQuestionId("")
-                          setUserId("")
+                          setStatus(null)
+                          setQuestionId(null)
+                          setUserId(null)
+                          setSuspend(null)
                         }}> Yes </button>
                       </div>
                   </div>
